HPC/requirements.txt
cupy-cuda11x>=12.0.0
numpy>=1.21.0

